requirements.txt
Flask==2.2.2
gunicorn==20.1.0
Jinja2==3.1.2
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==2.0.7
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6